Fix mapping of TileProvider.NO_TILE
When a TileProvider does not want to provide a tile for specific
coordinates, it returns NO_TILE. The mapping layer was passing on null
in that case, which is the signal for an error and will result in a
retry. The mapping layer was updated to use the GMS/HMS versions of
NO_TILE instead of null.