Fix latitude and longitude in LatLng
LatLng's constructors in the GMS and HMS implementations might change
the values for latitude and longitude but the wrapper library was always
using the original values passed to the constructor resulting in
different values in the mapper library and the maps implementations.

This fix uses the updated values of lat/lng from the GMS
implementation. The GMS and HMS implementations of LatLng are equivalent
and both wrapped objects should always be present.